# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory System API",
        "description": "Dokumentasi API untuk Manajemen Katalog Produk",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server Lokal"
        }
    ],
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Mengambil daftar produk dengan filter wajib",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Kategori produk (contoh: Elektronik, Pakaian)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "description": "Harga minimal",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil"
                    },
                    "422": {
                        "description": "Parameter wajib tidak diisi"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Menambah produk baru ke database",
                "operationId": "e8185e307706e2cd84bbcf5171e2f09d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "sku",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Kaos Oversize"
                                    },
                                    "sku": {
                                        "type": "string",
                                        "example": "TS-001"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 75000
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produk Berhasil Dibuat"
                    },
                    "422": {
                        "description": "Data Tidak Valid / SKU Sudah Ada"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Mengupdate data produk",
                "operationId": "6a7e11900844cd85cff66e662b82bd4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Produk yang ingin diupdate",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Hoodie Edit"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Clothing"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 150000
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update Berhasil"
                    },
                    "404": {
                        "description": "Produk Tidak Ditemukan"
                    },
                    "422": {
                        "description": "Validasi Gagal"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Menghapus produk berdasarkan ID",
                "operationId": "e80a6ca46716acee6b47ea1ac91663c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Produk yang ingin dihapus",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produk berhasil dihapus"
                    },
                    "404": {
                        "description": "Produk tidak ditemukan"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Products",
            "description": "Products"
        }
    ]
}